+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to this project will be documented in this file.
 
 * Add build on node 16.x
 * [FIX] Wait for module connection before configuring cached accessories
+* Support multiple cache files to allow more than 1 connection to Omni controller(s) via Child Bridges
 
 ## 1.4.0 (2021-04-16)