+++ CHANGELOG.md
@@ -5,10 +5,11 @@ All notable changes to this project will be documented in this file.
 ## NEXT
 
 * Add build on node 16.x
-* [FIX] Wait for module connection before configuring cached accessories
+* [FIX] Wait for Omni connection before configuring cached accessories
 * Support multiple cache files to allow more than 1 connection to Omni controller(s) via Child Bridges
 * Support Humidity control for Thermostat accessories
 * Minor improvements to some log messages (such as temperature display)
+* [FIX] Auto-Discovery not finding auxiliary sensors
 
 ## 1.4.0 (2021-04-16)